Melengkapkan mainEmployee dan perbaikan tiap class

+++ src/tugas/bab6/Employee.java
@@ -43,7 +43,7 @@ public class Employee { //superclassnya
         return tahunMasuk;
     }
 
-    public double getGaji() {
+    public int getGaji() {
         return gaji;
     }
 }
